requirements.txt
# Core Dependencies
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2
numpy>=1.23.1
scipy>=1.9.0
scikit-learn>=1.1.0

# Deep Learning Frameworks
pytorch-lightning>=1.7.1
timm>=0.9.12
einops>=0.7.0

# Computer Vision
opencv-python>=4.6.0
Pillow>=10.2.0
albumentations>=1.3.0
imgaug>=0.4.0

# Data Processing
pandas>=1.5.0
h5py>=3.11.0
pyyaml>=6.0.0
omegaconf>=2.3.0
pyrootutils>=1.0.4

# Visualization
matplotlib>=3.6.0
seaborn>=0.13.2
plotly>=5.15.0
tensorboard>=2.14.0

# Face Recognition Specific
insightface>=0.7.3
facenet-pytorch>=2.5.0
mtcnn>=0.1.1
retinaface-pytorch>=0.0.12

# Utilities
tqdm>=4.64.0
wandb>=0.16.6
tensorboard>=2.14.0
pyrootutils>=1.0.4
fvcore>=0.1.6

# Development Tools
pytest>=7.2.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
isort>=5.12.0

# Jupyter and Interactive Development
jupyter>=1.0.0
ipykernel>=6.29.4
notebook>=6.5.0

# Optional: For specific use cases
# mxnet>=1.9.1  # Uncomment if using MXNet models
# tensorflow>=2.10.0  # Uncomment if using TensorFlow models
